+++ vm-memory/src/volatile_memory.rs
@@ -34,10 +34,11 @@ use std::ptr::copy;
 use std::ptr::{read_volatile, write_volatile};
 use std::result;
 use std::sync::atomic::Ordering;
+use zerocopy::{CastError, FromBytes, FromZeros, Immutable, IntoBytes, KnownLayout};
 
 use crate::atomic_integer::AtomicInteger;
 use crate::bitmap::{Bitmap, BitmapSlice, BS};
-use crate::{AtomicAccess, ByteValued, Bytes};
+use crate::{AtomicAccess, Bytes};
 
 #[cfg(all(feature = "backend-mmap", feature = "xen", target_family = "unix"))]
 use crate::mmap::xen::{MmapXen as MmapInfo, MmapXenSlice};
@@ -131,7 +132,10 @@ pub trait VolatileMemory {
     }
 
     /// Gets a `VolatileRef` at `offset`.
-    fn get_ref<T: ByteValued>(&self, offset: usize) -> Result<VolatileRef<'_, T, BS<'_, Self::B>>> {
+    fn get_ref<T: Copy + Send + Sync + FromBytes + IntoBytes + FromZeros>(
+        &self,
+        offset: usize,
+    ) -> Result<VolatileRef<'_, T, BS<'_, Self::B>>> {
         let slice = self.get_slice(offset, size_of::<T>())?;
 
         assert_eq!(
@@ -156,7 +160,7 @@ pub trait VolatileMemory {
 
     /// Returns a [`VolatileArrayRef`](struct.VolatileArrayRef.html) of `n` elements starting at
     /// `offset`.
-    fn get_array_ref<T: ByteValued>(
+    fn get_array_ref<T: Copy + Send + Sync + FromBytes + IntoBytes + FromZeros>(
         &self,
         offset: usize,
         n: usize,
@@ -202,24 +206,36 @@ pub trait VolatileMemory {
     ///
     /// If the resulting pointer is not aligned, this method will return an
     /// [`Error`](enum.Error.html).
-    unsafe fn aligned_as_ref<T: ByteValued>(&self, offset: usize) -> Result<&T> {
+    unsafe fn aligned_as_ref<
+        T: Copy + Send + Sync + FromBytes + IntoBytes + FromZeros + Immutable + KnownLayout,
+    >(
+        &self,
+        offset: usize,
+    ) -> Result<&T> {
         let slice = self.get_slice(offset, size_of::<T>())?;
-        slice.check_alignment(align_of::<T>())?;
 
-        assert_eq!(
-            slice.len(),
-            size_of::<T>(),
-            "VolatileMemory::get_slice(offset, count) returned slice of length != count."
-        );
+        let slice_addr = slice.addr;
+        let size_of_t = size_of::<T>();
 
-        // SAFETY: This is safe because the invariants of the constructors of VolatileSlice ensure that
-        // slice.addr is valid memory of size slice.len(). The assert above ensures that
-        // the length of the slice is exactly enough to hold one `T`.
-        // Dereferencing the pointer is safe because we check the alignment above, and the invariants
-        // of this function ensure that no aliasing pointers exist. Lastly, the lifetime of the
-        // returned VolatileArrayRef match that of the VolatileSlice returned by get_slice and thus the
-        // lifetime one `self`.
-        unsafe { Ok(&*(slice.addr as *const T)) }
+        // SAFETY: Creating a byte slice from the pointer is safe because the invariants of the
+        // constructors of VolatileSlice ensure that slice.addr is valid memory of size slice.len()
+        // and the invariants of this function ensure that no aliasing reference exists.
+        // Lastly, the lifetime of the returned reference matches that of the VolatileSlice returned
+        // by get_slice and thus the lifetime one `self`.
+        let rust_slice = unsafe { core::slice::from_raw_parts(slice_addr, size_of_t) };
+        T::ref_from_bytes(rust_slice).map_err(|error| match error {
+            CastError::<_, T>::Alignment(_) => Error::Misaligned {
+                addr: slice.addr.addr(),
+                alignment: align_of::<T>(),
+            },
+            CastError::<_, T>::Size(size) => {
+                panic!("VolatileMemory::get_slice({}, count) returned slice of length != count, expected: \"{}\" actual: \"{}\".", offset, size_of::<T>(), size.into_src().len());
+            }
+            CastError::<_, T>::Validity(infallible) => {
+                // Ensures this error is actually infallible.
+                match infallible {}
+            }
+        })
     }
 
     /// Returns a mutable reference to an instance of `T` at `offset`. Mutable accesses performed
@@ -238,25 +254,36 @@ pub trait VolatileMemory {
     // the function is unsafe, and the conversion is safe if following the safety
     // instrutions above
     #[allow(clippy::mut_from_ref)]
-    unsafe fn aligned_as_mut<T: ByteValued>(&self, offset: usize) -> Result<&mut T> {
+    unsafe fn aligned_as_mut<
+        T: Copy + Send + Sync + FromBytes + IntoBytes + FromZeros + KnownLayout,
+    >(
+        &self,
+        offset: usize,
+    ) -> Result<&mut T> {
         let slice = self.get_slice(offset, size_of::<T>())?;
-        slice.check_alignment(align_of::<T>())?;
 
-        assert_eq!(
-            slice.len(),
-            size_of::<T>(),
-            "VolatileMemory::get_slice(offset, count) returned slice of length != count."
-        );
+        let slice_addr = slice.addr;
+        let size_of_t = size_of::<T>();
 
-        // SAFETY: This is safe because the invariants of the constructors of VolatileSlice ensure that
-        // slice.addr is valid memory of size slice.len(). The assert above ensures that
-        // the length of the slice is exactly enough to hold one `T`.
-        // Dereferencing the pointer is safe because we check the alignment above, and the invariants
-        // of this function ensure that no aliasing pointers exist. Lastly, the lifetime of the
-        // returned VolatileArrayRef match that of the VolatileSlice returned by get_slice and thus the
-        // lifetime one `self`.
-
-        unsafe { Ok(&mut *(slice.addr as *mut T)) }
+        // SAFETY: Creating a mutable byte slice from the pointer is safe because the invariants of the
+        // constructors of VolatileSlice ensure that slice.addr is valid memory of size slice.len()
+        // and the invariants of this function ensure that no aliasing reference exists.
+        // Lastly, the lifetime of the returned reference matches that of the VolatileSlice returned
+        // by get_slice and thus the lifetime one `self`.
+        let rust_slice = unsafe { core::slice::from_raw_parts_mut(slice_addr, size_of_t) };
+        T::mut_from_bytes(rust_slice).map_err(|error| match error {
+            CastError::<_, T>::Alignment(_) => Error::Misaligned {
+                addr: slice.addr.addr(),
+                alignment: align_of::<T>(),
+            },
+            CastError::<_, T>::Size(size) => {
+                panic!("VolatileMemory::get_slice({}, count) returned slice of length != count, expected: \"{}\" actual: \"{}\".", offset, size_of::<T>(), size.into_src().len());
+            }
+            CastError::<_, T>::Validity(infallible) => {
+                // Ensures this error is actually infallible.
+                match infallible {}
+            }
+        })
     }
 
     /// Returns a reference to an instance of `T` at `offset`. Mutable accesses performed
@@ -578,7 +605,7 @@ impl<'a, B: BitmapSlice> VolatileSlice<'a, B> {
     /// ```
     pub fn copy_to<T>(&self, buf: &mut [T]) -> usize
     where
-        T: ByteValued,
+        T: Copy + Send + Sync + FromBytes + IntoBytes + FromZeros,
     {
         // A fast path for u8/i8
         if size_of::<T>() == 1 {
@@ -657,7 +684,7 @@ impl<'a, B: BitmapSlice> VolatileSlice<'a, B> {
     /// ```
     pub fn copy_from<T>(&self, buf: &[T])
     where
-        T: ByteValued,
+        T: Copy + Send + Sync + FromBytes + IntoBytes + FromZeros,
     {
         // A fast path for u8/i8
         if size_of::<T>() == 1 {
@@ -897,7 +924,7 @@ pub struct VolatileRef<'a, T, B = ()> {
 
 impl<T> VolatileRef<'_, T, ()>
 where
-    T: ByteValued,
+    T: Copy + Send + Sync + FromBytes + IntoBytes + FromZeros,
 {
     /// Creates a [`VolatileRef`](struct.VolatileRef.html) to an instance of `T`.
     ///
@@ -915,7 +942,7 @@ where
 #[allow(clippy::len_without_is_empty)]
 impl<'a, T, B> VolatileRef<'a, T, B>
 where
-    T: ByteValued,
+    T: Copy + Send + Sync + FromBytes + IntoBytes + FromZeros,
     B: BitmapSlice,
 {
     /// Creates a [`VolatileRef`](struct.VolatileRef.html) to an instance of `T`, using the
@@ -1028,7 +1055,7 @@ pub struct VolatileArrayRef<'a, T, B = ()> {
 
 impl<T> VolatileArrayRef<'_, T>
 where
-    T: ByteValued,
+    T: Copy + Send + Sync + FromBytes + IntoBytes + FromZeros,
 {
     /// Creates a [`VolatileArrayRef`](struct.VolatileArrayRef.html) to an array of elements of
     /// type `T`.
@@ -1046,7 +1073,7 @@ where
 
 impl<'a, T, B> VolatileArrayRef<'a, T, B>
 where
-    T: ByteValued,
+    T: Copy + Send + Sync + FromBytes + IntoBytes + FromZeros,
     B: BitmapSlice,
 {
     /// Creates a [`VolatileArrayRef`](struct.VolatileArrayRef.html) to an array of elements of
@@ -2030,12 +2057,11 @@ mod tests {
 
     #[test]
     fn test_read_from_exceeds_size() {
-        #[derive(Debug, Default, Copy, Clone)]
+        #[derive(Debug, Default, Copy, Clone, FromBytes, IntoBytes)]
         struct BytesToRead {
             _val1: u128, // 16 bytes
             _val2: u128, // 16 bytes
         }
-        unsafe impl ByteValued for BytesToRead {}
         let cursor_size = 20;
         let image = vec![1u8; cursor_size];
 
@@ -2045,7 +2071,7 @@ mod tests {
         assert_eq!(
             image
                 .as_slice()
-                .read_volatile(&mut bytes_to_read.as_bytes())
+                .read_volatile(&mut VolatileSlice::from(bytes_to_read.as_mut_bytes()))
                 .unwrap(),
             cursor_size
         );
@@ -2268,7 +2294,7 @@ mod tests {
         index: usize,
         page_size: NonZeroUsize,
     ) where
-        T: ByteValued + From<u8>,
+        T: Copy + Send + Sync + FromBytes + IntoBytes + FromZeros + From<u8>,
     {
         let bitmap = AtomicBitmap::new(size_of_val(buf), page_size);
         let arr = unsafe {