+++ vm-memory/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Upcoming Release
+
+## Removed
+
+- \[[#52](https://github.com/rust-vmm/rust-vmm/pull/52)\] Removed `ByteValued` trait and `endian` module in favor of the `zerocopy` crate.
+  `ByteValued`'s functionality is now provided by `zerocopy::IntoBytes`, `zerocopy::FromBytes` and `zerocopy::Zeroed`.
+  `ByteValued::as_bytes` can be replaced with `VolatileSlices::from` for `&mut [u8]`.
+
 ## 0.18.0
 
 ### Changed